% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{items_static_metadata_full}
\alias{items_static_metadata_full}
\title{items_static_metadata_full.jl}
\usage{
items_static_metadata_full(path = "../datasets", force = FALSE)
}
\arguments{
\item{path}{(string) chosen dir to store the data}

\item{force}{(boolean) If file already exists, should it be downloaded again anyway?}
}
\value{
data.frame
}
\description{
Download and load of items_static_metadata_full.jl
}
\examples{

items_static_metadata_full <- items_static_metadata_full()

}
