% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{test_data}
\alias{test_data}
\title{test_data.csv}
\usage{
test_data(path = "../datasets", force = FALSE)
}
\arguments{
\item{path}{(string) chosen dir to store the data}

\item{force}{(boolean) If file already exists, should it be downloaded again anyway?}
}
\value{
data.frame
}
\description{
Download and load of test_data.csv
}
\examples{

test_data <- test_data()

}
