% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{train_data}
\alias{train_data}
\title{train_data.parquet}
\usage{
train_data(path = "../datasets", force = FALSE)
}
\arguments{
\item{path}{(string) chosen dir to store the data}

\item{force}{(boolean) If file already exists, should it be downloaded again anyway?}
}
\value{
data.frame
}
\description{
Download and load of train_data.parquet
}
\examples{

train_data <- train_data()

}
